creates game_logic.py, moves game logic into game_logic.py
